feat: 자동차의 상태를 출력.
- ResultView 화면출력 기능 추가.
- RandomGenerator Test 분리
