+++ src/main/java/com/megazone/goop/racingcar/logic/Race.java
@@ -2,16 +2,20 @@ package com.megazone.goop.racingcar.logic;
 
 import java.util.ArrayList;
 
+import com.megazone.goop.racingcar.ui.ResultView;
+
 public class Race {
 
     private static final int MAX_RANDOM_VALUE = 10;
 
+    private final ResultView resultView;
     private final RandomGenerator randomGenerator;
     private ArrayList<Car> cars;
 
     private int currentRound = 0;
 
-    public Race(RandomGenerator randomGenerator, int carCount) {
+    public Race(ResultView resultView, RandomGenerator randomGenerator, int carCount) {
+        this.resultView = resultView;
         this.randomGenerator = randomGenerator;
 
         joinCars(carCount);
@@ -20,6 +24,9 @@ public class Race {
     public void startRace(int roundCount) {
         for (int i = 0; i < roundCount; i++) {
             cars.forEach(c -> c.move(randomGenerator.getRandomValue(MAX_RANDOM_VALUE)));
+
+            resultView.displayResult(cars.stream().map(Car::getTotalDistance).toList());
+
             currentRound++;
         }
     }
