+++ src/test/java/com/megazone/goop/racingcar/logic/CarTest.java
@@ -1,6 +1,7 @@
 package com.megazone.goop.racingcar.logic;
 
-import org.junit.jupiter.api.*;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
@@ -8,24 +9,11 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 class CarTest {
 
-    RandomGenerator randomGenerator;
     Car car;
 
     @BeforeEach
     void setUp() {
-        randomGenerator = new RandomGenerator();
         car = new Car();
-    }
-
-    @DisplayName("매번 0~9 사이의 랜덤값을 구한다.")
-    @ParameterizedTest
-    @ValueSource(ints = {0, 1, 2, 3, 4, 5, 6, 7, 8, 9})
-    void randomBetweenTest(int value) {
-        int bound = 10;
-
-        assertThat(randomGenerator.getRandomValue(bound)).isGreaterThanOrEqualTo(0);
-
-        assertThat(randomGenerator.getRandomValue(bound)).isLessThanOrEqualTo(9);
     }
 
     @DisplayName("자동차는 랜덤값이 4 이상인 경우에 1 전진한다.")
